Add S3 creation flow draft 9 - rename the class of s3 configuration
